Fix PR #115 code review notes
